chore: apply PR suggestions
